+++ kwimage/im_filter.py
@@ -107,6 +107,7 @@ def fourier_mask(img_hwc, mask, axis=None, clip=None):
         xdoctest -m kwimage.im_filter fourier_mask --show
 
     Example:
+        >>> from kwimage.im_filter import *  # NOQA
         >>> import kwimage
         >>> img_hwc = kwimage.grab_test_image(space='gray')
         >>> mask = np.random.rand(*img_hwc.shape[0:2])